Add files via upload
Working game state implementation. We should be utilizing this for the rest of the changes made on the game. The vast majority of the game logic is handled in the "GameScreen" file, and everything else is more or less the same with the exception of the 'XScreen" files which control the menu system.