Added an AnimationHandler, the update testing page, and the new sprite sheet you'll need for the powerups
The AnimationHandler holds all of the animations we need access to for the game with all of the necessary getter methods. When you're using them just instantiate a new instance of AnimationHandler then treat the getter methods as you would a normal animation. Make sure to put the new sprite sheet in your data file or it definitely will not compile.